+++ pcbnew/autorouter/autorout.h
@@ -209,7 +209,7 @@ void TraceFilledRectangle( int ux0, int uy0, int ux1, int uy1,
 
 /* Same as above, but the rectangle is inclined angle angle. */
 void TraceFilledRectangle( int ux0, int uy0, int ux1, int uy1,
-                           int angle, LAYER_MSK masque_layer,
+                           double angle, LAYER_MSK masque_layer,
                            int color, int op_logic );
 
 /* QUEUE.CPP */
